requirements.txt
pycryptodome>=3.19.0
cryptography>=41.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
requests>=2.31.0
